Move log utils into navigation utils
