+++ frontend/amundsen_application/static/js/utils/index.spec.ts
@@ -3,7 +3,6 @@ import { ResourceType } from 'interfaces/Resources';
 import { UpdateMethod } from 'interfaces/Enums';
 
 import * as DateUtils from './dateUtils';
-import * as LogUtils from './logUtils';
 import * as NavigationUtils from './navigationUtils';
 import * as NumberUtils from './numberUtils';
 import * as TextUtils from './textUtils';
@@ -254,6 +253,42 @@ describe('navigationUtils', () => {
       expect(actual).toEqual(expected);
     });
   });
+
+  describe('getLoggingParams', () => {
+    let searchString;
+    let replaceStateSpy;
+
+    beforeAll(() => {
+      replaceStateSpy = jest.spyOn(window.history, 'replaceState');
+    });
+
+    it('returns the parsed source and index in an object', () => {
+      searchString = 'source=test_source&index=10';
+      const params = NavigationUtils.getLoggingParams(searchString);
+      expect(params.source).toEqual('test_source');
+      expect(params.index).toEqual('10');
+    });
+
+    it('clears the logging params from the URL, if present', () => {
+      const uri = 'testUri';
+      const mockFilter = '{"tag":"tagName"}';
+      searchString = `uri=${uri}&filters=${mockFilter}&source=test_source&index=10`;
+      replaceStateSpy.mockClear();
+      NavigationUtils.getLoggingParams(searchString);
+      expect(replaceStateSpy).toHaveBeenCalledWith(
+        {},
+        '',
+        `${window.location.origin}${window.location.pathname}?uri=${uri}&filters=${mockFilter}`
+      );
+    });
+
+    it('does not clear the logging params if they do not exist', () => {
+      searchString = '';
+      replaceStateSpy.mockClear();
+      NavigationUtils.getLoggingParams(searchString);
+      expect(replaceStateSpy).not.toHaveBeenCalled();
+    });
+  });
 });
 
 describe('dateUtils', () => {
@@ -304,44 +339,6 @@ describe('dateUtils', () => {
     const config = { timestamp: 1580421964000 };
     const dateString = DateUtils.formatDateTimeLong(config);
     expect(dateString).toEqual('January 30th 2020 at 10:06:04 pm');
-  });
-});
-
-describe('logUtils', () => {
-  describe('getLoggingParams', () => {
-    let searchString;
-    let replaceStateSpy;
-
-    beforeAll(() => {
-      replaceStateSpy = jest.spyOn(window.history, 'replaceState');
-    });
-
-    it('returns the parsed source and index in an object', () => {
-      searchString = 'source=test_source&index=10';
-      const params = LogUtils.getLoggingParams(searchString);
-      expect(params.source).toEqual('test_source');
-      expect(params.index).toEqual('10');
-    });
-
-    it('clears the logging params from the URL, if present', () => {
-      const uri = 'testUri';
-      const mockFilter = '{"tag":"tagName"}';
-      searchString = `uri=${uri}&filters=${mockFilter}&source=test_source&index=10`;
-      replaceStateSpy.mockClear();
-      LogUtils.getLoggingParams(searchString);
-      expect(replaceStateSpy).toHaveBeenCalledWith(
-        {},
-        '',
-        `${window.location.origin}${window.location.pathname}?uri=${uri}&filters=${mockFilter}`
-      );
-    });
-
-    it('does not clear the logging params if they do not exist', () => {
-      searchString = '';
-      replaceStateSpy.mockClear();
-      LogUtils.getLoggingParams(searchString);
-      expect(replaceStateSpy).not.toHaveBeenCalled();
-    });
   });
 });
